+++ include/common.hpp
@@ -18,7 +18,7 @@ namespace rft
    /// Size of the SHA256 hash
    const uint8_t SHA256_SIZE = 32;
    /// Maximum receiving window size of a client
-   const uint8_t MAX_WINDOW_SIZE = 64;
+   const uint16_t MAX_WINDOW_SIZE = 64;
 }// namespace rft
 // ------------------------------------------------------------------------
 #endif//ROBUST_FILE_TRANSFER_COMMON_HPP
